Show votes validators correctly

+++ types/src/committee.rs
@@ -9,7 +9,7 @@ use crate::model::ModelWithId;
 #[persistable(prefix = "/blocks/c")]
 pub struct CommitteeModel {
     pub index: u64,
-    pub validators: Vec<u64>,
+    pub validators: Vec<usize>,
 }
 
 #[cfg(feature = "indexing")]
@@ -17,7 +17,7 @@ impl From<&lighthouse_types::OwnedBeaconCommittee> for CommitteeModel {
     fn from(value: &lighthouse_types::OwnedBeaconCommittee) -> Self {
         CommitteeModel {
             index: value.index,
-            validators: value.committee.iter().map(|i| *i as u64).collect(),
+            validators: value.committee.iter().map(|i| *i).collect(),
         }
     }
 }
